Modulo 6/Clase5/App/requirements.txt
fastapi[all]
psycopg2-binary
sqlalchemy
alembic
pydantic
pytest
httpx
python-dotenv
pytest-asyncio